requirements.txt
chainlit
python-dotenv 